long added in estimated cost

+++ ERP_V2/src/main/java/com/example/ERP_V2/DTO/OrderDTO.java
@@ -50,5 +50,5 @@ public class OrderDTO {
 
     private BigDecimal plateRate;
 
-    private BigDecimal estimatedAmount;
+    private long estimatedAmount;
 }
